split beamline and utils tests
